Rename package activity as panels
