+++ app/src/androidTest/java/com/github/houseorganizer/houseorganizer/MainActivityTest.java
@@ -2,15 +2,13 @@ package com.github.houseorganizer.houseorganizer;
 
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.assertion.ViewAssertions.matches;
-import static androidx.test.espresso.intent.Intents.intended;
-import static androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
-import com.github.houseorganizer.houseorganizer.activity.MainActivity;
+import com.github.houseorganizer.houseorganizer.panels.MainActivity;
 
 import org.junit.FixMethodOrder;
 import org.junit.Rule;
