+++ app/src/main/java/com/github/houseorganizer/houseorganizer/panels/InfoActivity.java
@@ -1,4 +1,4 @@
-package com.github.houseorganizer.houseorganizer.activity;
+package com.github.houseorganizer.houseorganizer.panels;
 
 import android.content.SharedPreferences;
 import android.os.Bundle;
