+++ app/src/main/res/layout/activity_info.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.InfoActivity">
+    tools:context=".panels.InfoActivity">
 
     <TextView
         android:id="@+id/info_text_view"
