+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#673AB7"
-    tools:context=".activity.MainActivity">
+    tools:context=".panels.MainActivity">
 
     <ImageView
         android:id="@+id/house_logo"
